解了个ACTION_READ时file为null的bug，并将doc文档补充完整

+++ src/lab/sodino/util/NetworkUtil.java
@@ -209,7 +209,9 @@ public class NetworkUtil {
 					
 					if(read != contentLength){
 						info.resultCode = DOWNLOAD_DATA_LOSSY;
-						fileSaveTmp.delete();
+						if(info.dataAction == DownloadInfo.ACTION_SAVE && fileSaveTmp != null){
+							fileSaveTmp.delete();
+						}
 						LogOut.out(NetworkUtil.class.getName(), "DOWNLOAD_DATA_LOSSY result=" + info.resultCode +" url:" + urlString);
 						return;
 					} else {
